give 'test_db_migrates' 3 retries in ci to fix flakiness (#2518)
should fix #2402

any sqlite database created after september 12 2024 already has a plaintext header. the migration to plaintext would only effect clients created before that date. further, the error this test fails on seems tightly coupled to glibc and sqlicipher rather than something in libxmtp, and is difficult to reproduce locally. therefore it seems safe to simply give this test a few retries, esp as the test still passes, and is only failing b/c of a memory error in stderr.