Fix duplicate welcome error when two streams process the same welcome (#2548)
We have some reports of agents receiving the following error:

```
{"timestamp":"2025-10-01T22:20:16.994287Z","level":"WARN","message":"failed to create group from welcome due to duplicate welcome ID: storage error: the welcome id Some(137555076) already exists","target":"xmtp_mls::groups::welcome_sync"}
Caught error AgentError: Error occured during conversation streaming.
    at onError (/Users/bennycode/dev/bennycode/xmtp-siteassist-agent/node_modules/@xmtp/agent-sdk/src/core/Agent.ts:306:13)
    at streamCallback (/Users/bennycode/dev/bennycode/xmtp-siteassist-agent/node_modules/@xmtp/node-sdk/src/utils/streams.ts:106:14) {
  [cause]: [Error: storage error: the welcome id Some(137555076) already exists] {
    code: 'GenericFailure'
  }
}
```

From my digging through the code, this specific error variant could be thrown if we receive a welcome which is:
1. Greater than the sync cursor
2. We don't see an existing group when we check
3. By the time we write it, a corresponding group with the same welcome ID has already been stored in the database.

This is a race condition that should be possible if two different streams process the same welcome (neither of which updates the sync cursor), and both streams perform the check for an existing group before either stream has written it.

My fix is to catch the error and returning the `AlreadyProcessed` error code instead, which will cause the stream handling logic to swallow the error and fetch the existing DB entry. A riskier but more logical alternative would have been to move the check for an existing group (2) into the same transaction as the one which writes the group, which requires some refactoring.

Note that the existing group is only re-fetched when the welcome is processed by stream. If the welcome is processed by sync, this matches the existing behavior, where the cursor is incremented but the existing group is not returned. I suspect that is okay for now because the return value from `sync_welcomes()` is only used for tests, as far as I can tell.

I wrote a unit test (in `mls.rs`) but am unable to reproduce due to the race condition being so narrow, sadly.