api primitives compatible with v3 and d14n (#2507)
This is a PR broken out of the cursor migration to make that PR smaller

Creates some new primitives that are compatible with d14n and v3. in the process, adds some nice QoL changes to make working with the types easier. for instance, converts our timestamps to chrono types. the mls message will be deserialized in the api client so we can access is_commit, group id and other useful fields (req. for d14n to assign originators correctly, as well as only deser once). the actual usage/impl of the types are in upper PRs.

the idea here is to have version normalization happen lower in the stack then the xmtp_mls crate. these new types are free to change however necessary and adapt to protobuf versioning changes. the 'extraction' of a `group_message::V1` will now happen in xmtp_api_d14n instead of `xmtp_mls`. mls will just see these new primitives and the api will accept them.

using rusts strong typing also gives us the opportunity to enshrine more invariants into the codebase. for instance, if a type is an `InstallationId`/`GroupId` or `Cursor` we can make assumptions about it that would not be possible with a `Vec<u8>`. small improvements like this can snowball into easier-to read code further up the stack

this pr also introduces a `Generate` trait. in our tests, we have a frequent pattern of generating a dummy version of X struct and doing a bunch of stuff with it. the generate trait makes that pattern official and easier to use. dummy type can be generated with `T::generate()`. usage of this can be observed in the group/welcome message extractor, where we can generate real-form mls message so that the extractors can deserialize/use openmls functions.