Add readd intent (#2563)
This implements the publish flow for a `readd_installations` method (which will be called in upstack PR's).

Once published, the readd commit will fail current validation logic once received. Updating the validation logic (which is a breaking change) is tackled in the next two PR's. Therefore, the unit test currently asserts that the method will error.