Fix content type decoders (node/wasm) (#2935)
# Summary

The node and wasm bindings automatically decode `EncodedContent` when
converting from `StoredGroupMessage` to `Message`. The decode functions
of these bindings assumed that the raw decrypted bytes would be present
in `Message`, like the FFI bindings.

This PR updates the decode functions to accept the already decoded
`EncodedContent` in `Message` instead of the raw bytes.